feat(refact): dived the app in multiple files and add autocomplete
